+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - New classes have been added to support feed URL generation. See
   `\Lullabot\Mpx\Service\Feeds` for details. #111
 
+### Changed
+
+- The `\Lullabot\Mpx\Player\Url` has been moved to
+  `\Lullabot\Mpx\Service\Player\Url`. #111
+
 ## [0.6.2] - 2018-07-17
 
 ### Added